Isolate Onshape validation and diagnostics in MCP core
